--- app/definition/Versions.scala
@@ -1,34 +1,0 @@
-/*
- * Copyright 2022 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package definition
-
-import play.api.http.HeaderNames.ACCEPT
-import play.api.mvc.RequestHeader
-
-object Versionsx {
-  val VERSION_1 = "1.0"
-  val VERSION_2 = "2.0"
-
-  private val versionRegex = """application\/vnd.hmrc.(\d.\d)\+json""".r
-
-  def getFromRequest(request: RequestHeader): Option[String] =
-    getFrom(request.headers.headers)
-
-  private def getFrom(headers: Seq[(String, String)]): Option[String] =
-    headers.collectFirst { case (ACCEPT, versionRegex(ver)) => ver }
-
-}
